add the demo app.
edit the completion of AKView,fix bug, add the demo app.

+++ AKLabelView/AKCircleView.h
@@ -10,5 +10,5 @@
 
 @interface AKCircleView : AKView
 @property (nonatomic, assign) AKPoint originPoint;
-@property (nonatomic, assign) NSInteger radius;
+@property (nonatomic, assign) CGFloat radius;
 @end
